# server/routes/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Music Library Management API",
    "description": "API for managing music library integration with MusicBrainz, Lidarr, and OIDC authentication",
    "version": "1.0.0",
    "contact": {
      "name": "Music Library API Support"
    }
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "sessionAuth": []
    }
  ],
  "paths": {
...
    "/api/musicbrainz/recording": {
      "get": {
        "tags": ["MusicBrainz"],
        "summary": "Search recordings",
        "description": "Search MusicBrainz recordings database",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query string"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Recording search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using OIDC"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
    }
  },
...
}